backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.2
pymongo==4.10.1
flask-pymongo==2.3.0
flask-jwt-extended==4.5.3
bcrypt==4.0.1
werkzeug==2.3.7
